Fix Float kattype test (it was testing the Int type instead. :/)
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcp-python/trunk@3886 b56c7a7d-ac47-0410-a905-b6f7de78896a

+++ katcp/test/test_kattypes.py
@@ -74,11 +74,11 @@ class TestInt(TestType):
 class TestFloat(TestType):
 
     def setUp(self):
-        basic =  Int()
-        default = Int(default=11.0)
-        optional = Int(optional=True)
-        default_optional = Int(default=11.0, optional=True)
-        self.minmax = Int(min=5.0, max=6.0)
+        basic =  Float()
+        default = Float(default=11.0)
+        optional = Float(optional=True)
+        default_optional = Float(default=11.0, optional=True)
+        self.minmax = Float(min=5.0, max=6.0)
 
         self._pack = [
             (basic, 5.0, "5"),
